refactor: move addGlobalStyles into css-utils
Consolidate the shared CSS helpers by moving addGlobalStyles from
add-global-styles.js into css-utils.js and removing the old module.
