+++ packages/component-base/src/styles/style-props.js
@@ -4,8 +4,7 @@
  * This program is available under Apache License Version 2.0, available at https://vaadin.com/license/
  */
 import { css } from 'lit';
-import { registerCSSProperty } from '../css-utils.js';
-import { addGlobalStyles } from './add-global-styles.js';
+import { addGlobalStyles, registerCSSProperty } from '../css-utils.js';
 
 // NOTE: Base color CSS custom properties are explicitly registered as `<color>`
 // here to avoid performance issues in Aura. Aura overrides these properties with
