+++ packages/component-base/src/styles/user-colors.js
@@ -4,7 +4,7 @@
  * This program is available under Apache License Version 2.0, available at https://vaadin.com/license/
  */
 import { css } from 'lit';
-import { addGlobalStyles } from './add-global-styles.js';
+import { addGlobalStyles } from '../css-utils.js';
 
 addGlobalStyles(
   'vaadin-base-user-colors',
